Add snapshot tests for typeAware in nested configs.
Currently it does not seem to work as one file is missing a lint violation when it should have it. But I'm unsure whether that's because of a bug in the code or a misconfiguration of the test. Will maybe need help on that.

This file should have a diagnostic but does not: `crates/oxc_language_server/src/linter/snapshots/fixtures_linter_tsgolint_nested_type_aware@nested__test-with-no-lsp-config.ts.snap`

As an aside, it'd be preferable if I had a good way to add descriptions to the frontmatter of these snapshots to say what they should have in them. And it'd be nice if the snaps pointed back to the original rs test somehow. But this is fine for now, I guess.